Style all buttons consistently
- Remove interactive text
- Apply consistent font size, width, and animation to all buttons, navigation links, and recipe cards
